feat: improve logging and hide verbose messages by default (#554)
By default, now only ` ➜ Storybook: http://localhost:6006/ ` is printed,
similar to the devtools and tailwind module. The verbose output is
helpful for debugging, and can be activated via the module options (see
playground for an example). The build output from storybook is still
displayed, fixing this will be done in a follow-up.
Fixes #432.