fix: vite vue plugin order
`@vitejs/plugin-vue` should be the first registered user plugin so that it will be added directly after Vite's core plugins and transforms global vue compoments before nuxt:components:imports. The transform stack order can be reviewed with the `vite-plugin-inspect` plugin.

This fixes the use of global components in projects or modules.

Refs: #4