feat: add enabled option (#612)
Fixes #555 and fixes
https://github.com/storybook-vue/storybook-nuxt/issues/63.

This adds an `enabled` option. 

This can be useful for layers.
Some users have project A, B with A extending B that has storybook but
is not needed in A. This can improve dev server startup as storybook can
be quite heavy to load on bigger projects.