feat: Exposing the https options for Storybook dev server (#772)
### 📚 Description
With this module it's easy to set-up a storybook env. but it isn't
possible to use a ssl certificate like the CLI does:
`storybook dev -p 6006 --https --ssl-cert ./secrets/cert.pem --ssl-key
./secrets/key.pem`.

To make that possible with this module I've added/exposed the https
options. I used the same format as Nuxt
[devServer.https](https://nuxt.com/docs/api/nuxt-config#https) options.

You could even argue that we could just reuse the https options from the
devServer section directly if they're defined. But to keep it flexible
I've kept them separate.

### How to test
- Generate a SSL key & cert with e.g. https://github.com/Subash/mkcert
- Add those to the playground folder
- Add in the nuxt.config.ts of the playground with the right paths:
```
storybook: {
    https: {
      key: './key.pem',
      cert: './cert.pem',
    },
},
```
- `pnpm dev`
- See the storybook running on https://[...]

---------

Co-authored-by: Brandon de Vries <82502009+brandondv-NLO@users.noreply.github.com>