Fix: Extend the server proxy preset to also fetch SSG NuxtIslands (#815)
### 📚 Description
This PR extends the preset for the server proxy, to also allow the
fetching of NuxtIsland components from the dev server.

Standalone stories for NuxtIslands work fine. But when you have a
component that has a NuxtIsland component in it, it won't render
(properly) because it can't find the server only parts in the Storybook
server. With this PR those routes are exposed and make the stories
render.

There's one thing left to make NuxtIsland's work without any user
changes needed: a Suspense wrapper, but I wasn't too sure if/where that
should be enabled by default.

## Reproduction stories

https://github.com/nuxt-modules/storybook/compare/main...brandondv:nuxtjs-storybook:feature/child-nuxt-island-stories

Co-authored-by: Brandon de Vries <82502009+brandondv-NLO@users.noreply.github.com>