feat!: upgrade storybook to v9 (#882)
<!--
☝️ PR title should follow conventional commits
(https://conventionalcommits.org).
In particular, the title should start with one of the following types:

- docs: 📖 Documentation (updates to the documentation or readme)
- fix: 🐞 Bug fix (a non-breaking change that fixes an issue)
- feat: ✨ New feature/enhancement (a non-breaking change that adds
functionality or improves existing one)
- feat!/fix!: ⚠️ Breaking change (fix or feature that would cause
existing functionality to change)
- chore: 🧹 Chore (updates to the build process or auxiliary tools and
libraries)
-->

### 🔗 Linked issue

<!-- If it resolves an open issue, please link the issue here. For
example "Resolves #123" -->

### 📚 Description

Upgrade to v9 of storybook and also updates all examples.

This is a breaking change as it's not possible to support v8 and v9 at
the same time, as confirmed in
https://github.com/nuxt-modules/storybook/pull/882#issuecomment-2830073540.

Fixes https://github.com/storybookjs/storybook/issues/31608.

<!-- Describe your changes in detail -->
<!-- Why is this change required? What problem does it solve? -->

---------

Co-authored-by: Valentin Palkovic <dev@valentinpalkovic.dev>