fix: 🐞 Fix semver range update in Renovate in peerDependencies breaking peer dependency compatability (#910)
<!--
☝️ PR title should follow conventional commits
(https://conventionalcommits.org).
In particular, the title should start with one of the following types:

- docs: 📖 Documentation (updates to the documentation or readme)
- fix: 🐞 Bug fix (a non-breaking change that fixes an issue)
- feat: ✨ New feature/enhancement (a non-breaking change that adds
functionality or improves existing one)
- feat!/fix!: ⚠️ Breaking change (fix or feature that would cause
existing functionality to change)
- chore: 🧹 Chore (updates to the build process or auxiliary tools and
libraries)
-->

### 🔗 Linked issue

Resolves #907

### 📚 Description

This PR fixes peerDependency storyblock dependency to use a semver
range. In #882 storybook was implemented as a peer dependency to the 2
packages of this repo, and was given a range of `^9.0.4`. But then in
#905 the Renovate bot updated storybook to `9.0.5`, including the peer
dependency of it, and pinned the version to `9.0.5`, leading to #907
This PR fixes that

NB: Make sure that the new rule I've implemented in the Renovate config
matches the group settings you want