fix: include config dir in tsconfig (#926)
### 🔗 Linked issue
* #920 
<!-- If it resolves an open issue, please link the issue here. For
example "Resolves #123" -->

### 📚 Description
With https://github.com/nuxt/nuxt/pull/32601 merged, the related issue
will be resolved for projects using nuxt 4.

This will also improve/add support for the directory structure in nuxt 4
as the `.storybook` directory is outside the app/server directories and
needs to be added manually by this module.

~~If we also want this resolved for projects using nuxt 3, the PR will
either need to be backported (ideal) or we could use my package
nuxt-module-utils (I have not done so in this PR, I don't want to be
pushy with my own packages 😂).~~
(this is not relevant since the separated tsconfig structure is nuxt 4
only)
<!-- Describe your changes in detail -->
<!-- Why is this change required? What problem does it solve? -->

---------

Co-authored-by: Tobias Diez <code@tobiasdiez.de>