chore: update dependencies to latest minor/patch (#933)
### 📚 Description

I saw that most dependencies where out of date and thought it might be
easier to manually bump them to ensure things work instead of relying on
renovate.

I updated most things except for `storybook`, `chromatic`, and major
versions.

I also migrated some `pnpm` config that we had in `package.json` to
`pnpm-workspace.yaml` to improve consistency and avoid possible future
conflicts between duplicated settings.

----

Tests seem to be passing on my codespace so maybe this will be easier to
merge than #885.

The storybook window than opens when running the tests is not broken
anymore (it was showing errors when running the tests before the
updates). Only issue I see there is that the styles for `NuxtWelcome`
are broken although I remember seeing this before so maybe it is
unrelated to these changes.

The playwright tests are also passing but I still see some of them
showing a red cross on that step as they were doing before the updates.
Any clue why this happens? Do we perhaps need to increase/remove the
timeout?

<img width="1018" height="739" alt="Captura de pantalla 2025-08-07 a las
17 58 23"
src="https://github.com/user-attachments/assets/c66b77e1-f096-45de-9755-ee70a86b4932"
/>